+++ backend/cuda/include/vecpar/cuda/detail/host_memory.hpp
@@ -200,41 +200,37 @@ namespace vecpar::cuda {
                                            vecmem::host_memory_resource& mr,
                                            vecmem::vector<T>& data,
                                            Arguments... args)  {
-        int size = data.size();
-        // copy input data from host to device
-        auto data_buffer = copy.to(vecmem::get_data(data), d_mem, vecmem::copy::type::host_to_device);
-        auto data_view = vecmem::get_data(data_buffer);
+      size_t size = data.size();
+      // copy input data from host to device
+      auto data_buffer = copy.to(vecmem::get_data(data), d_mem,
+                                 vecmem::copy::type::host_to_device);
+      auto data_view = vecmem::get_data(data_buffer);
 
-        // allocate temp map result on host and copy to device
-        vecmem::vector<R> map_result(size, &mr);
-        auto map_result_buffer = copy.to(vecmem::get_data(map_result), d_mem, vecmem::copy::type::host_to_device);
-        auto map_result_view = vecmem::get_data(map_result_buffer);
+      // allocate temp map result on host and copy to device
+      vecmem::vector<R> map_result(size, &mr);
+      auto map_result_buffer = copy.to(vecmem::get_data(map_result), d_mem,
+                                       vecmem::copy::type::host_to_device);
+      auto map_result_view = vecmem::get_data(map_result_buffer);
 
-        internal::parallel_map(size,
-                               algorithm,
-                               map_result_view,
-                               data_view,
-                               args...);
+      internal::parallel_map(size, algorithm, map_result_view, data_view,
+                             args...);
 
-        // allocate result on host and device
-        vecmem::vector<R>* result = new vecmem::vector<R>(size, &mr);
-        auto result_buffer = copy.to(vecmem::get_data(*result), d_mem,
-                                     vecmem::copy::type::host_to_device);
-        auto result_view = vecmem::get_data(result_buffer);
+      // allocate result on host and device
+      vecmem::vector<R> *result = new vecmem::vector<R>(size, &mr);
+      auto result_buffer = copy.to(vecmem::get_data(*result), d_mem,
+                                   vecmem::copy::type::host_to_device);
+      auto result_view = vecmem::get_data(result_buffer);
 
-        int* idx; // global index
-        cudaMallocManaged((void**)&idx, sizeof(int));
-        *idx = 0;
-        internal::parallel_filter(size,
-                                  algorithm,
-                                  idx,
-                                  result_view,
-                                  map_result_view);
+      int *idx; // global index
+      cudaMallocManaged((void **)&idx, sizeof(int));
+      *idx = 0;
+      internal::parallel_filter(size, algorithm, idx, result_view,
+                                map_result_view);
 
-        //copy back results
-        copy(result_buffer, *result, vecmem::copy::type::device_to_host);
-        result->resize(*idx);
-        return result;
+      // copy back results
+      copy(result_buffer, *result, vecmem::copy::type::device_to_host);
+      result->resize(*idx);
+      return result;
     }
 
     template<class Algorithm,
@@ -246,12 +242,13 @@ namespace vecpar::cuda {
                                            vecmem::host_memory_resource& mr,
                                            vecmem::vector<T>& data,
                                            Arguments... args)  {
-        int size = data.size();
-        // copy input data from host to device
-        auto data_buffer = copy.to(vecmem::get_data(data), d_mem, vecmem::copy::type::host_to_device);
-        auto data_view = vecmem::get_data(data_buffer);
+      size_t size = data.size();
+      // copy input data from host to device
+      auto data_buffer = copy.to(vecmem::get_data(data), d_mem,
+                                 vecmem::copy::type::host_to_device);
+      auto data_view = vecmem::get_data(data_buffer);
 
-        // allocate temp map result on host and copy to device
+      // allocate temp map result on host and copy to device
       //  vecmem::vector<R> map_result(data.size(), &mr);
       //  auto map_result_buffer = copy.to(vecmem::get_data(map_result), d_mem, vecmem::copy::type::host_to_device);
       //  auto map_result_view = vecmem::get_data(map_result_buffer);
