+++ backend/cuda/include/vecpar/cuda/detail/kernels.hpp
@@ -5,25 +5,20 @@
 
 namespace vecpar::cuda {
 
-    template<typename Function, typename... Arguments>
-    __global__ void kernel(int size,
-                           Function f,
-                           Arguments... args) {
-        int idx = blockIdx.x * blockDim.x + threadIdx.x;
-        if (idx >= size)
-            return;
-        f(idx, args...);
-    }
+template <typename Function, typename... Arguments>
+__global__ void kernel(size_t size, Function f, Arguments... args) {
+  size_t idx = blockIdx.x * blockDim.x + threadIdx.x;
+  if (idx >= size)
+    return;
+  f(idx, args...);
+}
 
-    template<typename Function, typename... Arguments>
-    __global__ void rkernel(int *lock,
-                            int size,
-                            Function f,
-                            Arguments... args) {
-        int idx = blockIdx.x * blockDim.x + threadIdx.x;
-        if (idx >= size)
-            return;
-        f(lock, args...);
-    }
+template <typename Function, typename... Arguments>
+__global__ void rkernel(int *lock, size_t size, Function f, Arguments... args) {
+  size_t idx = blockIdx.x * blockDim.x + threadIdx.x;
+  if (idx >= size)
+    return;
+  f(lock, args...);
+}
 }
 #endif //VECPAR_KERNELS_HPP
